+++ packages/webcomponents/src/components/form/form-control-text.tsx
@@ -44,6 +44,7 @@ export class TextInput {
     const target = event.target;
     const name = target.getAttribute('name');
     this.inputHandler(name, target.value);
+    this.formControlInput.emit(target.value);
   }
 
   componentDidLoad() {
@@ -61,9 +62,7 @@ export class TextInput {
           name={this.name}
           onInput={(event: any) => this.handleFormControlInput(event)}
           onBlur={() => this.formControlBlur.emit()}
-          part={`input ${this.error && 'input-invalid'} ${
-            this.disabled && 'input-disabled'
-          }`}
+          part={`input ${this.error ? 'input-invalid ' : ''}${this.disabled ? ' input-disabled' : ''}`}
           class={this.error ? 'form-control is-invalid' : 'form-control'}
           type="text"
           disabled={this.disabled}